requirements.txt
Django==5.0.1
python-dotenv==1.0.0
cohere>=5.1.4,<6.0.0
pinecone-client==3.0.3
cloudinary==1.38.0
PyPDF2==3.0.1
python-docx==1.1.0
pandas==2.2.0
Pillow==10.2.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
